Move compliance_loader to core/compliance/loaders.py
